Refactor room to include room ID in title
