+++ app/src/app/room/[roomId]/page.tsx
@@ -1,45 +1,17 @@
-'use client'
+import RoomClient from '@/components/page/room/RoomClient/RoomClient'
+import { DefaultSearchParams, ServerParams } from '@/types/Params'
+import { Metadata } from 'next'
 
-import InRoom from '@/components/page/room/InRoom/InRoom'
-import LoadingRoom from '@/components/page/room/LoadingRoom/LoadingRoom'
-import { useAuthContext } from '@/context/Auth'
-import { useRoom } from '@/context/Room'
-import { useSocket } from '@/context/Socket'
-import { useParams, useRouter } from 'next/navigation'
-import { useEffect } from 'react'
+export async function generateMetadata(
+	props: ServerParams<DefaultSearchParams, { roomId: string }>
+): Promise<Metadata> {
+	const { roomId } = await props.params
 
-export default function Room() {
-	const { room, joinRoom } = useRoom()
-	const { roomId } = useParams<{ roomId: string }>()
-	const router = useRouter()
-	const socket = useSocket()
-	const auth = useAuthContext()
-
-	useEffect(() => {
-		if (
-			socket.state === 'loading' ||
-			auth.state === 'loading' ||
-			auth.state === 'reloading' ||
-			room
-		) {
-			return
-		}
-		if (socket.state === 'error' || auth.state === 'error') {
-			router.push('/')
-			return
-		}
-
-		init()
-	}, [socket, auth, room])
-
-	async function init() {
-		const res = await joinRoom(roomId)
-		if (res.error) return router.push('/')
+	return {
+		title: `Room ${roomId}`
 	}
+}
 
-	if (!room) {
-		return <LoadingRoom />
-	}
-
-	return <InRoom />
+export default function Page() {
+	return <RoomClient />
 }
